refactor: move `ftl initdb` into its own command `ftl-initdb`
This is to avoid an issue where the DB couldn't be initialised if the
CLI was in a state where it couldn't be built, which happens quite
frequently during development.
