+++ backend/controller/databasetesting/devel.go
@@ -1,4 +1,4 @@
-package sql
+package databasetesting
 
 import (
 	"context"
