+++ scripts/ftl-initdb
@@ -1,0 +1,1 @@
+ftl
